pyproject.toml
[tool.poetry]
name = "abm_logistics"
version = "0.1.0"
description = ""
authors = ["artush.ghazaryan96 <artush.ghazaryan96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
psycopg2 = "^2.9.1"
djangorestframework = "^3.12.4"
django-rest-marshmallow = "^4.0.2"
django-rest-swagger = "^2.2.0"
marshmallow = "^3.12.1"
drf-yasg = "^1.20.0"
Pillow = "^8.3.0"
django-rest-auth = "^0.9.5"
django-allauth = "^0.44.0"
djangorestframework-simplejwt = "^4.7.2"
django-cors-headers = "^3.7.0"
gunicorn = "^20.1.0"
supervisor = "^4.2.2"
django-modeltranslation = "^0.17.3"
django-currentuser = "^0.5.3"
django-simple-history = "^3.0.0"
celery = "^5.1.2"
redis = "^3.5.3"
flower = "^1.0.0"
django-celery-beat = "^2.2.1"
django-celery-results = "^2.2.0"
fcm-django = "^1.0.4"
channels = "^3.0.4"
channels-redis = "^3.3.0"
uvicorn = "^0.15.0"
websockets = "^9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
